Rename package to @crawlee/fs-storage
