+++ docs/guides/result_storage.mdx
@@ -8,7 +8,7 @@ import ApiLink from '@site/src/components/ApiLink';
 
 Crawlee has several result storage types that are useful for specific tasks. The data is stored on a local disk to the directory defined by the `CRAWLEE_STORAGE_DIR` environment variable. If this variable is not defined, by default Crawlee sets `CRAWLEE_STORAGE_DIR` to `./storage` in the current working directory.
 
-By default, Crawlee storage is managed by the <ApiLink to="filesystem-storage/class/FileSystemStorageClient">`FileSystemStorageClient`</ApiLink> class, which stores all information as local files in the respective storage type folders.
+By default, Crawlee storage is managed by the <ApiLink to="fs-storage/class/FileSystemStorageClient">`FileSystemStorageClient`</ApiLink> class, which stores all information as local files in the respective storage type folders.
 
 ## Key-value store
