+++ packages/core/package.json
@@ -53,7 +53,7 @@
         "@apify/pseudo_url": "^2.0.59",
         "@apify/timeout": "^0.3.2",
         "@apify/utilities": "^2.15.5",
-        "@crawlee/filesystem-storage": "workspace:*",
+        "@crawlee/fs-storage": "workspace:*",
         "@crawlee/memory-storage": "workspace:*",
         "@crawlee/types": "workspace:*",
         "@crawlee/utils": "workspace:*",
