+++ packages/crawlee/package.json
@@ -54,7 +54,7 @@
         "@crawlee/cheerio": "workspace:*",
         "@crawlee/cli": "workspace:*",
         "@crawlee/core": "workspace:*",
-        "@crawlee/filesystem-storage": "workspace:*",
+        "@crawlee/fs-storage": "workspace:*",
         "@crawlee/http": "workspace:*",
         "@crawlee/jsdom": "workspace:*",
         "@crawlee/linkedom": "workspace:*",
