+++ packages/fs-storage/.npmignore
