+++ packages/fs-storage/src/background-handler/fs-utils.ts
