+++ packages/fs-storage/src/background-handler/index.ts
