+++ packages/fs-storage/src/resource-clients/common/base-client.ts
