+++ packages/fs-storage/src/utils.ts
