+++ packages/fs-storage/test/fs-fallback.test.ts
@@ -2,7 +2,7 @@ import { randomUUID } from 'node:crypto';
 import { rm, writeFile } from 'node:fs/promises';
 import { resolve } from 'node:path';
 
-import { FileSystemStorageClient } from '@crawlee/filesystem-storage';
+import { FileSystemStorageClient } from '@crawlee/fs-storage';
 import type { KeyValueStoreRecord } from '@crawlee/types';
 import { ensureDir } from 'fs-extra/esm';
