+++ packages/fs-storage/test/no-crash-on-big-buffers.test.ts
@@ -4,7 +4,7 @@
 import { rm } from 'node:fs/promises';
 import { resolve } from 'node:path';
 
-import { FileSystemStorageClient } from '@crawlee/filesystem-storage';
+import { FileSystemStorageClient } from '@crawlee/fs-storage';
 import type { KeyValueStoreClient } from '@crawlee/types';
 
 describe('FileSystemStorageClient should not crash when saving a big buffer', () => {
