+++ pnpm-workspace.yaml
@@ -1,27 +1,27 @@
 packages:
-    - packages/*
-    - docs
-    - website
+  - packages/*
+  - docs
+  - website
 
 minimumReleaseAge: 1440
 minimumReleaseAgeExclude:
-    - "@apify/*"
-    - "@crawlee/*"
-    - "apify-client"
-    - "apify"
-    - "crawlee"
-    - "got-scraping"
+  - "@apify/*"
+  - "@crawlee/*"
+  - "apify-client"
+  - "apify"
+  - "crawlee"
+  - "got-scraping"
 
 overrides:
-    playwright-core: 1.60.0
-    "@puppeteer/browsers": ^3.0.4
-    "@browserbasehq/stagehand": 3.0.7
-    # Dedup minimatch to v9 everywhere except inside lerna — lerna 9.x's bundled
-    # code (`__toESM(require('minimatch')).default(...)`) only works with v3,
-    # whose CJS export *is* the function. Pinning v9 there caused the publish
-    # step to silently report 0 changed packages.
-    minimatch: "^9.0.0"
-    "lerna>minimatch": "^3.1.4"
+  playwright-core: 1.60.0
+  "@puppeteer/browsers": ^3.0.4
+  "@browserbasehq/stagehand": 3.0.7
+  # Dedup minimatch to v9 everywhere except inside lerna — lerna 9.x's bundled
+  # code (`__toESM(require('minimatch')).default(...)`) only works with v3,
+  # whose CJS export *is* the function. Pinning v9 there caused the publish
+  # step to silently report 0 changed packages.
+  minimatch: "^9.0.0"
+  "lerna>minimatch": "^3.1.4"
 
 # pnpm 11 replaces `onlyBuiltDependencies` with an explicit `allowBuilds` map.
 # Each entry must be true (build allowed) or false (build skipped) — pnpm 11
@@ -29,19 +29,20 @@ overrides:
 # (combined with `strictDepBuilds: false` below so the install can still
 # proceed when new build-requesting deps appear without a manual entry).
 allowBuilds:
-    "@apify/ui-icons": true
-    "@playwright/browser-chromium": true
-    "@playwright/browser-firefox": true
-    "@playwright/browser-webkit": true
-    "@swc/core": true
-    better-sqlite3: true
-    bufferutil: true
-    core-js: true
-    esbuild: true
-    nx: true
-    protobufjs: true
-    puppeteer: true
-    unrs-resolver: true
+  "@apify/ui-icons": true
+  "@playwright/browser-chromium": true
+  "@playwright/browser-firefox": true
+  "@playwright/browser-webkit": true
+  "@swc/core": true
+  better-sqlite3: true
+  bufferutil: true
+  core-js: true
+  core-js-pure: set this to true or false
+  esbuild: true
+  nx: true
+  protobufjs: true
+  puppeteer: true
+  unrs-resolver: true
 
 strictDepBuilds: false
 
@@ -60,7 +61,7 @@ nodeLinker: hoisted
 linkWorkspacePackages: true
 preferWorkspacePackages: true
 publicHoistPattern:
-    - "*"
+  - "*"
 
 patchedDependencies:
-    "@docusaurus/core@3.9.2": patches/@docusaurus__core@3.9.2.patch
+  "@docusaurus/core@3.9.2": patches/@docusaurus__core@3.9.2.patch
