+++ website/docusaurus.config.js
@@ -16,7 +16,7 @@ const packages = [
     'linkedom-crawler',
     'stagehand-crawler',
     'memory-storage',
-    'filesystem-storage',
+    'fs-storage',
     'utils',
     'types',
     'impit-client',
@@ -35,7 +35,7 @@ const packagesOrder = [
     '@crawlee/http-client',
     '@crawlee/browser',
     '@crawlee/memory-storage',
-    '@crawlee/filesystem-storage',
+    '@crawlee/fs-storage',
     '@crawlee/browser-pool',
     '@crawlee/utils',
     '@crawlee/types',
